refactor : 토큰 발급의 책임을 다른 클래스로 위임하도록 변경 #7
